requirements.txt
tqdm==4.60.0
numpy==1.20.2
PubChemPy==1.0.4
tensorflow==2.3.0
jsonlines==2.0.0
ChemSpiPy==2.0.0
pandas==1.2.3
transformers==4.11.3
ray==1.4.1
gym==0.18.0
torch==1.8.1
matplotlib==3.4.2
mat2vec==0.2
requests==2.25.1
Keras==2.4.3
gensim==4.0.1
nltk==3.6.1
beautifulsoup4==4.10.0
python_Levenshtein==0.12.2
rdkit==2009.Q1-1
scikit_learn==1.0.2
tokenizers==0.7.0
